Enable parallel tests by moving includes to one file
